requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
joblib==1.4.2
kombu==5.5.1
numpy==1.23.0
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
